fix: Resolve Member type conversion issue in CustomAuthenticationFilter
- Fixed issue where a Member object was being converted to a String
- Used appropriate fields (e.g., member.getUsername()) when storing JWT claims
- Resolved type mismatch issue when using Map<String, String>
- Removed unnecessary type casting to prevent compilation errors

+++ NextSpring/src/main/java/com/study/nextspring/global/base/UtClass.java
@@ -48,8 +48,7 @@ public class UtClass {
     }
 
     public static class jwt {
-        public static String toString(String secret, int expireSeconds, Map<String, Object> body) {
-
+        public static String toString(String secret, long expireSeconds, Map<String, Object> body) {
             Date issuedAt = new Date();
             Date expiration = new Date(issuedAt.getTime() + 1000L * expireSeconds);
